+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ### Added
 - `foreColor` & `hiliteColor`
 - Add @1x resources to solve the problem that some user resources are not found
+- Add `onCursorPosition` Enter the position of the cursor
 
 #1.7.0
 ### Added